Refactoring: Move all remotes into a seperate python file
This makes sense because:

* This file has no dependencies beyond the standard Python libraries,
  unlike stbt.py which depends on GStreamer, OpenCV, etc.
* The interface between this file and stbt.py is very narrow: it's
  limited to `uri_to_remote` and `uri_to_remote_recorder`.

but I'm doing it so I can reuse the remotes in fake-lircd which will
allow proxying the use of any of these remotes over a socket.
